+++ example/pages/components/index.js
@@ -1,26 +1,25 @@
 import React, { Component } from 'react';
-import Head from 'next/head';
 import Router from 'next/router';
-import { PrismCode } from 'react-prism';
-import { Table } from 'reactstrap';
 import Layout from '../../components/ComponentsLayout';
 
 
 export default class PlayerPage extends Component {
-  static async getInitialProps (ctx) {
-    const redirectToPage = '/components/player'
-    if (ctx && ctx.req) {
-      ctx.res.writeHead(301, {
-        Location: redirectToPage
-      })
-      ctx.res.end()
+  static async getInitialProps(ctx) {
+    const redirectToPage = '/components/player';
+    if (ctx && ctx.res) {
+      if (ctx.res.writeHead) {
+        ctx.res.writeHead(301, {
+          Location: redirectToPage,
+        });
+        ctx.res.end();
+      }
     } else {
-      Router.push(redirectToPage)
+      Router.push(redirectToPage);
     }
+    return {};
   }
 
   render() {
-    const { PlayerExampleSource, PlayerControlExampleSource } = this.props;
     return (
       <Layout page="Home">
         <h3>Components</h3>