+++ example/pages/customize/index.js
@@ -1,18 +1,22 @@
 import React, { Component } from 'react';
 import Head from 'next/head';
+import Router from 'next/router';
 import Layout from '../../components/CustomizeLayout';
 
-export default class EnableDisableComponentPage extends React.Component {
-  static async getInitialProps (ctx) {
-    const redirectToPage = '/customize/enable-disable-components'
-    if (ctx && ctx.req) {
-      ctx.res.writeHead(301, {
-        Location: redirectToPage
-      })
-      ctx.res.end()
+export default class EnableDisableComponentPage extends Component {
+  static async getInitialProps(ctx) {
+    const redirectToPage = '/customize/enable-disable-components';
+    if (ctx && ctx.res) {
+      if (ctx.res.writeHead) {
+        ctx.res.writeHead(301, {
+          Location: redirectToPage,
+        });
+        ctx.res.end();
+      }
     } else {
-      Router.push(redirectToPage)
+      Router.push(redirectToPage);
     }
+    return {};
   }
 
   render() {